create common action POOL
