+++ frontend/common/actions/index.js
@@ -1,1 +1,2 @@
 export * from './alerts';
+export * from './pool';
