+++ frontend/common/actions/pool.js
@@ -1,0 +1,6 @@
+export const POOL = {
+  RECEIVE_INFORMATION: 'pool/RECEIVE_INFORMATION',
+  RECEIVE_MESSAGES: 'pool/RECEIVE_MESSAGES',
+  RECEIVE_BULLETINS: 'pool/RECEIVE_BULLETINS',
+  CLEAR: 'pool/CLEAR'
+};
