+++ frontend/modules/pool/actions.js
@@ -1,14 +1,7 @@
-import { receiveAlerts } from 'common/actions';
+import { receiveAlerts, POOL } from 'common/actions';
 import { processMessages } from 'helpers';
 
 import * as PoolAPI from './utils';
-
-export const POOL = {
-  RECEIVE_INFORMATION: 'pool/RECEIVE_INFORMATION',
-  RECEIVE_MESSAGES: 'pool/RECEIVE_MESSAGES',
-  RECEIVE_BULLETINS: 'pool/RECEIVE_BULLETINS',
-  CLEAR: 'pool/CLEAR'
-};
 
 export const receivePoolInformation = pool => ({
   type: POOL.RECEIVE_INFORMATION,
